+++ apple-codesign/src/macho_signing.rs
@@ -598,7 +598,14 @@ impl<'data> MachOSigner<'data> {
                 .to_string(),
         );
 
+        // Team should only be included when signing with an Apple signed
+        // certificate. This logic is handled in [SigningSettings]. But emit
+        // a warning if the constraint is violated.
         let team_name = settings.team_id().map(|x| Cow::Owned(x.to_string()));
+
+        if team_name.is_some() && settings.signing_certificate_apple_signed() {
+            warn!("signing without an Apple signed certificate but signing settings contain a team name; signature varies from Apple's tooling");
+        }
 
         let mut cd = CodeDirectoryBlob {
             flags,